Closes: https://github.com/polkadot-fellows/runtimes/issues/436

~~(requested for `coretime-polkadot` and `polkadot`)~~

<!-- Remember that you can run `/merge` to enable auto-merge in the PR -->

<!-- Remember to modify the changelog. If you don't need to modify it, you can check the following box.
Instead, if you have already modified it, simply delete the following line. -->

- [X] Does not require a CHANGELOG entry

## TODO

- [x] add `polkadot-runtime` ~~(in-progress / running)~~
- [x] add `asset-hubs` ~~(in-progress / running)~~
- [x] add all
- [x] add subweight comparison 

## The comparison of this branch with `main@ab0f42df90ae60a7e6192afce842eb0341b17efa`

_Note: `Change [%]` has threashold 5, so it show only if change is more than `5%` or less than `-5%`._

# Relay chains

### Kusama

```
subweight compare commits          --path-pattern "./relay/kusama/**/weights/**/*.rs"          --format markdown --no-color           --change added changed          --method asymptotic --ignore-errors --strip-path-prefix="relay/kusama/src/weights/"          remotes/polkadot-fellows/main          origin/bko-weights
```
| File                                       | Extrinsic                               | Old      | New      | Change [%] |
|--------------------------------------------|-----------------------------------------|----------|----------|------------|
| polkadot_runtime_common_paras_registrar.rs | schedule_code_upgrade                   | -        | -        | ERROR      |
| runtime_parachains_paras.rs                | add_trusted_validation_code             | -        | -        | ERROR      |
| runtime_parachains_paras.rs                | force_note_new_head                     | -        | -        | ERROR      |
| runtime_parachains_paras.rs                | force_schedule_code_upgrade             | -        | -        | ERROR      |
| runtime_parachains_paras.rs                | force_set_current_code                  | -        | -        | ERROR      |
| runtime_parachains_paras.rs                | force_set_current_head                  | -        | -        | ERROR      |
| pallet_balances_nis_counterpart.rs         | burn_keep_alive                         | 19.23us  | 146.23us | +660.55    |
| pallet_balances_nis_counterpart.rs         | burn_allow_death                        | 28.44us  | 156.84us | +451.40    |
| runtime_parachains_assigner_on_demand.rs   | place_order_keep_alive                  | 333.04us | 735.26us | +120.77    |
| runtime_parachains_assigner_on_demand.rs   | place_order_allow_death                 | 332.99us | 735.00us | +120.73    |
| pallet_xcm.rs                              | execute                                 | 7.23us   | 12.10us  | +67.43     |
| runtime_parachains_coretime.rs             | request_revenue_at                      | 487.64us | 769.09us | +57.72     |
| frame_election_provider_support.rs         | phragmen                                | 40.47ms  | 58.87ms  | +45.44     |
| pallet_scheduler.rs                        | cancel                                  | 266.63us | 374.16us | +40.33     |
| pallet_staking.rs                          | cancel_deferred_slash                   | 5.88ms   | 8.22ms   | +39.82     |
| pallet_scheduler.rs                        | execute_dispatch_unsigned               | 1.90us   | 2.57us   | +35.62     |
| pallet_scheduler.rs                        | cancel_named                            | 299.30us | 405.78us | +35.57     |
| pallet_scheduler.rs                        | execute_dispatch_signed                 | 1.96us   | 2.62us   | +33.74     |
| pallet_referenda_fellowship_referenda.rs   | place_decision_deposit_not_queued       | 491.08us | 640.28us | +30.38     |
| pallet_referenda_fellowship_referenda.rs   | place_decision_deposit_queued           | 492.22us | 641.76us | +30.38     |
| frame_election_provider_support.rs         | phragmms                                | 35.71ms  | 46.50ms  | +30.23     |
| pallet_referenda_referenda.rs              | place_decision_deposit_preparing        | 417.11us | 528.42us | +26.69     |
| pallet_referenda_fellowship_referenda.rs   | cancel                                  | 405.70us | 513.85us | +26.66     |
| pallet_referenda_fellowship_referenda.rs   | place_decision_deposit_preparing        | 418.92us | 530.46us | +26.63     |
| pallet_referenda_referenda.rs              | cancel                                  | 405.09us | 512.64us | +26.55     |
| pallet_timestamp.rs                        | on_finalize                             | 3.40us   | 4.27us   | +25.44     |
| frame_system.rs                            | apply_authorized_upgrade                | 96.45ms  | 120.70ms | +25.14     |
| frame_system.rs                            | set_code                                | 94.74ms  | 118.45ms | +25.02     |
| pallet_referenda_fellowship_referenda.rs   | kill                                    | 462.30us | 573.67us | +24.09     |
| pallet_referenda_referenda.rs              | kill                                    | 482.67us | 596.14us | +23.51     |
| pallet_referenda_referenda.rs              | place_decision_deposit_queued           | 455.71us | 562.72us | +23.48     |
| pallet_xcm.rs                              | force_default_xcm_version               | 2.06us   | 2.54us   | +23.42     |
| pallet_referenda_referenda.rs              | place_decision_deposit_not_queued       | 455.96us | 562.58us | +23.38     |
| pallet_utility.rs                          | as_derivative                           | 3.98us   | 4.86us   | +22.11     |
| pallet_conviction_voting.rs                | remove_vote                             | 635.09us | 770.07us | +21.25     |
| pallet_scheduler.rs                        | service_task_base                       | 3.24us   | 3.91us   | +20.68     |
| pallet_referenda_referenda.rs              | place_decision_deposit_passing          | 551.01us | 663.45us | +20.41     |
| pallet_referenda_referenda.rs              | place_decision_deposit_failing          | 549.32us | 661.22us | +20.37     |
| frame_benchmarking_baseline.rs             | sr25519_verification                    | 4.06ms   | 4.87ms   | +20.14     |
| pallet_referenda_fellowship_referenda.rs   | place_decision_deposit_failing          | 580.23us | 693.49us | +19.52     |
| pallet_ranked_collective.rs                | vote                                    | 564.94us | 675.19us | +19.52     |
| pallet_referenda_fellowship_referenda.rs   | place_decision_deposit_passing          | 697.90us | 833.71us | +19.46     |
| pallet_scheduler.rs                        | service_task_periodic                   | 3.31us   | 3.94us   | +19.14     |
| frame_benchmarking_baseline.rs             | hashing                                 | 23.52ms  | 27.97ms  | +18.89     |
| pallet_staking.rs                          | deprecate_controller_batch              | 1.90s    | 2.25s    | +18.49     |
| pallet_utility.rs                          | dispatch_as                             | 6.19us   | 7.33us   | +18.47     |
| pallet_staking.rs                          | withdraw_unbonded_update                | 369.60us | 436.16us | +18.01     |
| pallet_fast_unstake.rs                     | on_idle_unstake                         | 48.76ms  | 56.97ms  | +16.83     |
| pallet_staking.rs                          | set_staking_configs_all_remove          | 603.31us | 704.72us | +16.81     |
| pallet_staking.rs                          | set_staking_configs_all_set             | 603.71us | 705.18us | +16.81     |
| pallet_balances_nis_counterpart.rs         | force_adjust_total_issuance             | 5.77us   | 6.71us   | +16.25     |
| pallet_utility.rs                          | batch                                   | 3.07ms   | 3.56ms   | +16.05     |
| pallet_conviction_voting.rs                | vote_existing                           | 933.51us | 1.08ms   | +15.93     |
| pallet_utility.rs                          | batch_all                               | 3.27ms   | 3.77ms   | +15.52     |
| frame_system.rs                            | remark_with_event                       | 6.08ms   | 7.02ms   | +15.40     |
| pallet_scheduler.rs                        | set_retry_named                         | 162.80us | 187.58us | +15.22     |
| frame_system.rs                            | authorize_upgrade                       | 108.23us | 124.63us | +15.16     |
| pallet_balances_native.rs                  | burn_allow_death                        | 28.44us  | 32.72us  | +15.03     |
| pallet_utility.rs                          | force_batch                             | 3.07ms   | 3.53ms   | +15.03     |
| pallet_election_provider_multi_phase.rs    | feasibility_check                       | 12.22ms  | 14.04ms  | +14.94     |
| pallet_scheduler.rs                        | set_retry                               | 134.05us | 153.40us | +14.44     |
| pallet_scheduler.rs                        | cancel_retry_named                      | 162.80us | 186.28us | +14.42     |
| pallet_preimage.rs                         | note_preimage                           | 8.82ms   | 10.06ms  | +14.03     |
| pallet_balances_native.rs                  | force_adjust_total_issuance             | 5.87us   | 6.67us   | +13.65     |
| pallet_scheduler.rs                        | cancel_retry                            | 134.05us | 152.12us | +13.48     |
| pallet_referenda_fellowship_referenda.rs   | nudge_referendum_requeued_slide         | 193.57us | 218.47us | +12.86     |
| pallet_balances_native.rs                  | burn_keep_alive                         | 19.23us  | 21.69us  | +12.81     |
| pallet_staking.rs                          | bond                                    | 515.83us | 581.78us | +12.79     |
| runtime_parachains_paras_inherent.rs       | enter_backed_candidates_variable        | 11.82ms  | 13.31ms  | +12.67     |
| runtime_parachains_paras_inherent.rs       | enter_variable_disputes                 | 11.18ms  | 12.59ms  | +12.66     |
| pallet_referenda_fellowship_referenda.rs   | nudge_referendum_requeued_insertion     | 193.71us | 218.15us | +12.62     |
| pallet_fast_unstake.rs                     | on_idle_check                           | 224.52ms | 252.34ms | +12.39     |
| xcm/pallet_xcm_benchmarks_generic.rs       | transact                                | 7.19us   | 8.02us   | +11.54     |
| pallet_referenda_fellowship_referenda.rs   | nudge_referendum_continue_confirming    | 342.59us | 380.91us | +11.18     |
| pallet_nomination_pools.rs                 | pool_withdraw_unbonded                  | 546.71us | 607.09us | +11.04     |
| pallet_fast_unstake.rs                     | deregister                              | 361.74us | 401.26us | +10.92     |
| pallet_conviction_voting.rs                | unlock                                  | 495.23us | 547.49us | +10.55     |
| pallet_nomination_pools.rs                 | chill                                   | 785.99us | 861.28us | +9.58      |
| pallet_referenda_fellowship_referenda.rs   | one_fewer_deciding_passing              | 482.90us | 526.11us | +8.95      |
| pallet_staking.rs                          | set_controller                          | 371.05us | 404.19us | +8.93      |
| pallet_conviction_voting.rs                | undelegate                              | 75.18ms  | 81.88ms  | +8.92      |
| pallet_conviction_voting.rs                | delegate                                | 75.50ms  | 81.95ms  | +8.55      |
| pallet_referenda_fellowship_referenda.rs   | nudge_referendum_end_confirming         | 409.51us | 444.46us | +8.54      |
| pallet_referenda_fellowship_referenda.rs   | one_fewer_deciding_failing              | 484.16us | 525.28us | +8.49      |
| pallet_referenda_fellowship_referenda.rs   | nudge_referendum_begin_deciding_passing | 475.36us | 515.35us | +8.41      |
| pallet_election_provider_multi_phase.rs    | submit_unsigned                         | 15.75ms  | 17.06ms  | +8.29      |
| pallet_preimage.rs                         | request_unnoted_preimage                | 164.31us | 177.87us | +8.25      |
| pallet_referenda_fellowship_referenda.rs   | nudge_referendum_queued                 | 326.20us | 352.82us | +8.16      |
| pallet_staking.rs                          | payout_stakers_alive_staked             | 255.62ms | 275.68ms | +7.85      |
| pallet_conviction_voting.rs                | remove_other_vote                       | 226.57us | 243.65us | +7.54      |
| pallet_nomination_pools.rs                 | nominate                                | 1.50ms   | 1.61ms   | +7.44      |
| pallet_preimage.rs                         | request_preimage                        | 174.50us | 186.86us | +7.08      |
| pallet_referenda_fellowship_referenda.rs   | nudge_referendum_not_queued             | 324.72us | 347.51us | +7.02      |
| pallet_staking.rs                          | restore_ledger                          | 566.21us | 604.99us | +6.85      |
| pallet_recovery.rs                         | as_recovered                            | 33.47us  | 35.68us  | +6.61      |
| pallet_proxy.rs                            | proxy                                   | 38.04us  | 40.46us  | +6.37      |
| pallet_nomination_pools.rs                 | withdraw_unbonded_kill                  | 2.83ms   | 3.00ms   | +5.94      |
| pallet_referenda_fellowship_referenda.rs   | nudge_referendum_rejected               | 407.27us | 431.07us | +5.85      |
| pallet_scheduler.rs                        | service_task_fetched                    | 5.29ms   | 5.59ms   | +5.75      |
| runtime_parachains_paras_inherent.rs       | enter_backed_candidate_code_upgrade     | 41.04ms  | 43.34ms  | +5.60      |
| pallet_nomination_pools.rs                 | withdraw_unbonded_update                | 1.31ms   | 1.38ms   | +5.58      |
| polkadot_runtime_common_paras_registrar.rs | set_current_head                        | 949.51us | 1.00ms   | +5.52      |
| runtime_parachains_hrmp.rs                 | hrmp_close_channel                      | 564.18us | 595.20us | +5.50      |
| pallet_xcm.rs                              | reserve_transfer_assets                 | 532.85us | 561.88us | +5.45      |
| pallet_preimage.rs                         | request_no_deposit_preimage             | 164.44us | 173.37us | +5.43      |
| runtime_parachains_hrmp.rs                 | hrmp_accept_open_channel                | 563.23us | 593.67us | +5.40      |
| runtime_parachains_hrmp.rs                 | establish_system_channel                | 1.16ms   | 1.22ms   | +5.22      |
| runtime_parachains_hrmp.rs                 | force_open_hrmp_channel                 | 1.17ms   | 1.23ms   | +5.22      |
| pallet_xcm.rs                              | claim_assets                            | 160.01us | 168.19us | +5.11      |
| runtime_parachains_inclusion.rs            | receive_upward_messages                 | 204.55ms | 191.47ms | -6.40      |
| pallet_election_provider_multi_phase.rs    | elect_queued                            | 1.98ms   | 1.86ms   | -6.47      |
| pallet_offences.rs                         | report_offence_grandpa                  | 11.12ms  | 10.34ms  | -6.99      |
| pallet_offences.rs                         | report_offence_babe                     | 11.13ms  | 10.34ms  | -7.02      |
| frame_benchmarking_baseline.rs             | subtraction                             | 209.86ns | 191.13ns | -8.92      |
| frame_benchmarking_baseline.rs             | addition                                | 210.33ns | 191.14ns | -9.12      |
| frame_benchmarking_baseline.rs             | multiplication                          | 208.95ns | 186.91ns | -10.55     |
| frame_benchmarking_baseline.rs             | division                                | 210.65ns | 187.04ns | -11.21     |
| xcm/pallet_xcm_benchmarks_generic.rs       | burn_asset                              | 1.73us   | 1.51us   | -12.46     |
| xcm/pallet_xcm_benchmarks_generic.rs       | refund_surplus                          | 1.88us   | 1.62us   | -13.92     |
| pallet_whitelist.rs                        | dispatch_whitelisted_call               | 6.14ms   | 5.26ms   | -14.40     |
| runtime_parachains_initializer.rs          | force_approve                           | 272.73us | 208.12us | -23.69     |
| runtime_parachains_disputes_slashing.rs    | report_dispute_lost                     | 2.42ms   | 1.79ms   | -26.08     |
| xcm/pallet_xcm_benchmarks_generic.rs       | expect_transact_status                  | 1.52us   | 1.12us   | -26.36     |
| xcm/pallet_xcm_benchmarks_generic.rs       | clear_topic                             | 1.27us   | 920.00ns | -27.62     |
| xcm/pallet_xcm_benchmarks_generic.rs       | set_appendix                            | 1.32us   | 950.00ns | -28.19     |
| xcm/pallet_xcm_benchmarks_generic.rs       | unpaid_execution                        | 1.34us   | 960.00ns | -28.36     |
| xcm/pallet_xcm_benchmarks_generic.rs       | expect_origin                           | 1.31us   | 930.00ns | -29.28     |
| xcm/pallet_xcm_benchmarks_generic.rs       | expect_error                            | 1.32us   | 930.00ns | -29.33     |
| xcm/pallet_xcm_benchmarks_generic.rs       | set_error_handler                       | 1.36us   | 950.00ns | -30.35     |
| xcm/pallet_xcm_benchmarks_generic.rs       | set_topic                               | 1.28us   | 890.00ns | -30.47     |
| xcm/pallet_xcm_benchmarks_generic.rs       | expect_asset                            | 1.50us   | 1.03us   | -31.29     |
| xcm/pallet_xcm_benchmarks_generic.rs       | clear_transact_status                   | 1.40us   | 960.00ns | -31.38     |
| xcm/pallet_xcm_benchmarks_generic.rs       | clear_error                             | 1.30us   | 890.00ns | -31.70     |
| xcm/pallet_xcm_benchmarks_generic.rs       | trap                                    | 1.31us   | 890.00ns | -32.16     |
| pallet_treasury.rs                         | on_initialize_proposals                 | 40.88ms  | 27.55ms  | -32.60     |
| xcm/pallet_xcm_benchmarks_generic.rs       | clear_origin                            | 1.31us   | 880.00ns | -33.08     |
| xcm/pallet_xcm_benchmarks_generic.rs       | set_fees_mode                           | 1.32us   | 880.00ns | -33.18     |
| xcm/pallet_xcm_benchmarks_generic.rs       | descend_origin                          | 1.40us   | 920.00ns | -34.24     |
| xcm/pallet_xcm_benchmarks_generic.rs       | buy_execution                           | 1.42us   | 920.00ns | -34.98     |
| pallet_scheduler.rs                        | schedule_retry                          | 466.87us | 237.40us | -49.15     |
| runtime_parachains_paras_inherent.rs       | enter_bitfields                         | 2.86ms   | 1.28ms   | -55.08     |
| pallet_nomination_pools.rs                 | apply_slash_fail                        | 234.68us | 3.32us   | -98.59     |
| pallet_nomination_pools.rs                 | apply_slash                             | 726.99us | 3.23us   | -99.56     |
| pallet_nomination_pools.rs                 | migrate_delegation                      | 1.13ms   | 3.05us   | -99.73     |
| pallet_nomination_pools.rs                 | pool_migrate                            | 1.65ms   | 3.11us   | -99.81     |

### Polkadot

```
subweight compare commits          --path-pattern "./relay/polkadot/**/weights/**/*.rs"          --format markdown --no-color           --change added changed          --method asymptotic --ignore-errors --strip-path-prefix="relay/polkadot/src/weights/"          remotes/polkadot-fellows/main          origin/bko-weights 
```
…