@bkontur 2024-08-15
those big numbers here:
```
| /pallet_broker.rs                      | request_revenue_info_at   | 240.00ns | 368.90us | +153608.33 |
| /pallet_broker.rs                      | notify_revenue            | 7.00us   | 102.58us | +1365.43   |
| /pallet_broker.rs                      | on_new_timeslice          | 7.00us   | 30.57us  | +336.71    |
```
looks ok, because previous values were just some "random" weights:

request_revenue_info_at:
```
Weight::from_parts(240_000, 0)
			.saturating_add(Weight::from_parts(0, 0))
```

notify_revenue:
```
Weight::from_parts(7_000_000, 7000)
```

on_new_timeslice:
```
Weight::from_parts(7_000_000, 7000)
```