@bkontur 2024-08-15
> > requested for coretime-polkadot and polkadot
> 
> We should do for all runtimes as part of the release though?

Donal wanted coretime-polkadot and polkadot and I wanted AssetHubs/BridgeHubs.
But yes, I am running the rest now