@alindima 2024-08-16
@bkontur can you give some info about the errors reported for `runtime_parachains_paras.rs`?