@bkontur 2024-08-16
> @bkontur can you give some info about the errors reported for `runtime_parachains_paras.rs`?

hmm, no idea, I just changed subweight [`method`](https://github.com/ggwpez/substrate-weight-compare?tab=readme-ov-file#evaluation-method) to `base` for ERROR-ed comparisions, because `asymptotic` fails.

@ggwpez Do you have idea why `asymptotic` fails here? I have never seen this before.