@bkontur 2024-08-16
`ERROR` from description with changed `method` from `asymptotic` to `base` and changed `threshold` from 5 to 1


## Kusama

```
subweight compare commits --threshold 1         --path-pattern "./relay/kusama/**/weights/**/runtime_parachains_paras.rs"          --format markdown --no-color           --change added changed          --method base --ignore-errors --strip-path-prefix="relay/kusama/src/weights/"          remotes/polkadot-fellows/main          origin/bko-weights
```
| File                        | Extrinsic                                              | Old      | New      | Change [%] |
|-----------------------------|--------------------------------------------------------|----------|----------|------------|
| runtime_parachains_paras.rs | force_queue_action                                     | 167.17us | 170.67us | +2.10      |
| runtime_parachains_paras.rs | include_pvf_check_statement_finalize_onboarding_reject | 276.21us | 281.82us | +2.03      |
| runtime_parachains_paras.rs | include_pvf_check_statement_finalize_upgrade_reject    | 278.61us | 283.79us | +1.86      |
| runtime_parachains_paras.rs | include_pvf_check_statement                            | 282.18us | 286.15us | +1.41      |
| runtime_parachains_paras.rs | add_trusted_validation_code                            | 470.59us | 409.35us | -13.01     |
| runtime_parachains_paras.rs | force_set_current_code                                 | 1.01ms   | 860.80us | -14.89     |

```
subweight compare commits --threshold 1         --path-pattern "./relay/kusama/**/weights/**/polkadot_runtime_common_paras_registrar.rs"          --format markdown --no-color           --change added changed          --method base --ignore-errors --strip-path-prefix="relay/kusama/src/weights/"          remotes/polkadot-fellows/main          origin/bko-weights
```
| File                                       | Extrinsic      | Old    | New    | Change [%] |
|--------------------------------------------|----------------|--------|--------|------------|
| polkadot_runtime_common_paras_registrar.rs | force_register | 8.30ms | 8.18ms | -1.37      |
| polkadot_runtime_common_paras_registrar.rs | swap           | 1.16ms | 1.14ms | -1.51      |

## Polkadot

```
subweight compare commits --threshold 1         --path-pattern "./relay/polkadot/**/weights/**/runtime_parachains_paras.rs"          --format markdown --no-color           --change added changed          --method base --ignore-errors --strip-path-prefix="relay/polkadot/src/weights/"          remotes/polkadot-fellows/main          origin/bko-weights
```
| File                        | Extrinsic                                              | Old      | New      | Change [%] |
|-----------------------------|--------------------------------------------------------|----------|----------|------------|
| runtime_parachains_paras.rs | include_pvf_check_statement_finalize_onboarding_accept | 988.76us | 1.03ms   | +4.29      |
| runtime_parachains_paras.rs | force_schedule_code_upgrade                            | 926.49us | 948.75us | +2.40      |
| runtime_parachains_paras.rs | force_queue_action                                     | 167.36us | 170.74us | +2.02      |
| runtime_parachains_paras.rs | add_trusted_validation_code                            | 471.43us | 476.98us | +1.18      |
| runtime_parachains_paras.rs | include_pvf_check_statement                            | 283.84us | 287.09us | +1.14      |
| runtime_parachains_paras.rs | force_set_current_code                                 | 899.93us | 861.24us | -4.30      |

```
subweight compare commits --threshold 1         --path-pattern "./relay/polkadot/**/weights/**/polkadot_runtime_common_paras_registrar.rs"          --format markdown --no-color           --change added changed          --method base --ignore-errors --strip-path-prefix="relay/polkadot/src/weights/"          remotes/polkadot-fellows/main          origin/bko-weights
```
| File                                       | Extrinsic             | Old      | New      | Change [%] |
|--------------------------------------------|-----------------------|----------|----------|------------|
| polkadot_runtime_common_paras_registrar.rs | swap                  | 1.11ms   | 1.51ms   | +35.16     |
| polkadot_runtime_common_paras_registrar.rs | deregister            | 601.92us | 641.71us | +6.61      |
| polkadot_runtime_common_paras_registrar.rs | register              | 8.09ms   | 8.50ms   | +5.00      |
| polkadot_runtime_common_paras_registrar.rs | force_register        | 8.12ms   | 8.47ms   | +4.34      |
| polkadot_runtime_common_paras_registrar.rs | reserve               | 299.77us | 306.19us | +2.14      |
| polkadot_runtime_common_paras_registrar.rs | schedule_code_upgrade | 1.25ms   | 963.65us | -23.04     |


## BridgeHubKusama

```
subweight compare commits --threshold 1          --path-pattern "./system-parachains/bridge-hubs/bridge-hub-kusama/**/weights/**/pallet_bridge_grandpa.rs"          --format markdown --no-color           --change added changed          --method base --ignore-errors --strip-path-prefix="system-parachains/bridge-hubs/bridge-hub-kusama/src/weights"          remotes/polkadot-fellows/main          origin/bko-weights
```
| File                      | Extrinsic              | Old      | New      | Change [%] |
|---------------------------|------------------------|----------|----------|------------|
| /pallet_bridge_grandpa.rs | force_set_pallet_state | 753.75us | 741.44us | -1.63      |
| /pallet_bridge_grandpa.rs | submit_finality_proof  | 985.05us | 870.35us | -11.64     |

## BridgeHubPolkadot

```
subweight compare commits --threshold 1         --path-pattern "./system-parachains/bridge-hubs/bridge-hub-polkadot/**/weights/**/pallet_bridge_grandpa.rs"          --format markdown --no-color           --change added changed          --method base --ignore-errors --strip-path-prefix="system-parachains/bridge-hubs/bridge-hub-polkadot/src/weights"          remotes/polkadot-fellows/main          origin/bko-weights
```
| File                      | Extrinsic              | Old      | New      | Change [%] |
|---------------------------|------------------------|----------|----------|------------|
| /pallet_bridge_grandpa.rs | submit_finality_proof  | 843.87us | 871.87us | +3.32      |
| /pallet_bridge_grandpa.rs | force_set_pallet_state | 753.75us | 742.80us | -1.45      |